requirements.txt
Django>=5.2.5
python-docx>=1.1.0
PyMuPDF>=1.24.0
